refactor: 수익 계산 기능 위치 수정
- 수익 계산을 State에서 하도록 수정
- Result클래스 제거, 승패 체크는 Stand 상태에서만 하도록 수정

+++ src/main/java/blackjack/dto/ProfitDto.java
@@ -3,9 +3,9 @@ package blackjack.dto;
 public class ProfitDto {
 
     private final String name;
-    private final int money;
+    private final double money;
 
-    public ProfitDto(String name, int money) {
+    public ProfitDto(String name, double money) {
         this.name = name;
         this.money = money;
     }
@@ -14,7 +14,7 @@ public class ProfitDto {
         return name;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 }
